:art: chore : 케밥 메뉴 토글 및 수정 중 수정하기 버튼 가리기 구현

+++ src/components/AnswerDelete/index.jsx
@@ -4,10 +4,12 @@ import PropTypes from 'prop-types';
 import { useState } from 'react';
 import { useNavigate } from 'react-router-dom';
 
-const AnswerDelete = ({ answerId, onAnswerDeleted }) => {
+const AnswerDelete = ({ id, answerId, onAnswerDeleted, onKebabClick }) => {
   AnswerDelete.propTypes = {
+    id: PropTypes.number.isRequired,
     answerId: PropTypes.number.isRequired,
     onAnswerDeleted: PropTypes.func.isRequired,
+    onKebabClick: PropTypes.func.isRequired,
   };
 
   const [isDeleting, setIsDeleting] = useState(false);
@@ -15,18 +17,22 @@ const AnswerDelete = ({ answerId, onAnswerDeleted }) => {
   const navigate = useNavigate();
 
   const handleAnswerDelete = async () => {
-    setIsDeleting(true);
+    const userConfirmed = window.confirm('정말로 삭제하시겠습니까?', ''); // user 확인작업은 confirm으로 임시로 만들었습니다.
+    if (userConfirmed) {
+      onKebabClick(id);
+      setIsDeleting(true);
 
-    try {
-      const response = await deleteAnswer(answerId);
-      if (!response.ok) {
-        throw new Error('답변 삭제 중 오류가 발생했습니다.');
+      try {
+        const response = await deleteAnswer(answerId);
+        if (!response.ok) {
+          throw new Error('답변 삭제 중 오류가 발생했습니다.');
+        }
+        onAnswerDeleted(answerId);
+      } catch (err) {
+        navigate('/');
+      } finally {
+        setIsDeleting(false);
       }
-      onAnswerDeleted(answerId);
-    } catch (err) {
-      navigate('/');
-    } finally {
-      setIsDeleting(false);
     }
   };
